create a base class for generator models
